Artificial Intellegence/APIs/requirements.txt
Flask
langchain
langchain-groq
python-dotenv
jsonschema
pydantic